BSPs: Client-side support for newer map formats (BSP2, 29a)

+++ gl_model.h
@@ -77,7 +77,7 @@ typedef struct texture_s {
 
 // !!! if this is changed, it must be changed in asm_draw.h too !!!
 typedef struct medge_s {
-	unsigned short		v[2];
+	unsigned int		v[2];
 	unsigned int		cachededgeoffset;
 } medge_t;
 
@@ -144,8 +144,8 @@ typedef struct mnode_s {
 	mplane_t			*plane;
 	struct mnode_s		*children[2];	
 
-	unsigned short		firstsurface;
-	unsigned short		numsurfaces;
+	unsigned int		firstsurface;
+	unsigned int		numsurfaces;
 } mnode_t;
 
 typedef struct mleaf_s {
